+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.23.10
 
 - Support for `--sandbox-runner` and `--dartdoc-command` CLI options.
--  Updated platform report: partial score for wasm and switft manager incompatibility.
+- Updated platform report: partial score for WASM and Swift package manager incompatibility.
+- Better parsing of `dart format` output.
 
 ## 0.23.9